refactor: 💡 수정 예시 코드 위치 변경
병합에 따른 파일명 변경으로 인해 AccountIdAware의 사용법을 담은 예시코드가 StoreRegisterController로 변경되었습니다.
